Basic rotation done and flipped rendering
