+++ tdraw/render.go
@@ -116,7 +116,15 @@ func TDraw(buff *image.RGBA, zbuff [][]float64, vew, nrm, tex Triangle, textureD
 	x1 := int(math.Max(vew.A.X(), math.Max(vew.B.X(), vew.C.X())))
 	y1 := int(math.Max(vew.A.Y(), math.Max(vew.B.Y(), vew.C.Y())))
 	dims := textureData.Bounds()
-	buffH := buff.Bounds().Max.Y
+
+	// If the drawn pixels are out of the screen's range then skip them entirely.
+	if (x0 < 0) || (x1 > len(zbuff)-1) {
+		return
+	}
+	if (y0 < 0) || (y1 > len(zbuff[0])-1) {
+		return
+	}
+
 	for x := x0; x <= x1; x++ {
 		for y := y0; y <= y1; y++ {
 			bc := vew.BaryCenter(x, y)
@@ -139,7 +147,7 @@ func TDraw(buff *image.RGBA, zbuff [][]float64, vew, nrm, tex Triangle, textureD
 					// Change from the original gel: we subtract y from buffH because,
 					// somewhere, I (200sc) messed up the translation and we accidentally
 					// are rendering everything upsidedown. This is the easiest fix!
-					buff.Set(x, buffH-y, PShade(textureData.At(int(xx), int(yy)), shading))
+					buff.Set(x, y, PShade(textureData.At(int(xx), int(yy)), shading))
 				}
 			}
 		}
